# backend/src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Incident Management API",
    "version": "1.0.0",
    "description": "API để quản lý báo cáo sự cố - Hệ thống báo cáo sự cố"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Root endpoint",
        "tags": ["General"],
        "responses": {
          "200": {
            "description": "API information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "version": { "type": "string" },
                    "status": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ping": {
      "get": {
        "summary": "Ping endpoint để kiểm tra kết nối",
        "tags": ["General"],
        "responses": {
          "200": {
            "description": "Backend is reachable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "timestamp": { "type": "string" },
                    "origin": { "type": "string" },
                    "corsAllowed": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Health check",
        "tags": ["General"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/audit-logs": {
      "get": {
        "summary": "Get audit logs",
        "tags": ["Audit"],
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Filter by action type"
          }
        ],
        "responses": {
          "200": {
            "description": "List of audit logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "array", "items": { "type": "object" } },
                    "count": { "type": "integer" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents": {
      "get": {
        "summary": "Get all incidents",
        "tags": ["Incidents"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Filter by status"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Filter by incident type"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of incidents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "type": "object" }
                    },
                    "count": { "type": "integer" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/upload-image": {
      "post": {
        "summary": "Upload image for an incident",
        "tags": ["Incidents"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file (max 5MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "imageUrl": { "type": "string" },
                    "filename": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/quick-report": {
      "post": {
        "summary": "Quick report (create incident)",
        "tags": ["Incidents"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["imageUrl", "type", "location", "timestamp"],
                "properties": {
                  "imageUrl": { "type": "string" },
                  "type": { "type": "string" },
                  "location": {
                    "type": "object",
                    "properties": {
                      "lat": { "type": "number" },
                      "lng": { "type": "number" },
                      "address": { "type": "string" }
                    }
                  },
                  "description": { "type": "string" },
                  "timestamp": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/stats/overview": {
      "get": {
        "summary": "Get incident statistics overview",
        "tags": ["Incidents"],
        "responses": {
          "200": {
            "description": "Statistics overview",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "pending": { "type": "integer" },
                        "resolved": { "type": "integer" },
                        "inProgress": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/recent": {
      "get": {
        "summary": "Get recent incidents",
        "tags": ["Incidents"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 10 },
            "description": "Number of recent incidents to return"
          }
        ],
        "responses": {
          "200": {
            "description": "List of recent incidents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "type": "object" }
                    },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{id}": {
      "get": {
        "summary": "Get incident by ID",
        "tags": ["Incidents"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Incident ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Incident object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/incidents/{id}/status": {
      "put": {
        "summary": "Update incident status",
        "tags": ["Incidents"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Incident ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["pending", "in_progress", "resolved", "rejected"]
                  },
                  "note": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}